+++ server/run/test_run.py
@@ -81,6 +81,16 @@ def run_all_cases(target_exec_dir, target, kind_bench, bench_name,
         os.remove(logfile)
 
     starttime = datetime.datetime.now()
+    if os.path.exists(Folder.caliper_log_file):
+        sections = bench_name + " EXECUTION"
+        fp = open(Folder.caliper_log_file,"r")
+        f = fp.readlines()
+        fp.close()
+        op = open(Folder.caliper_log_file,"w")
+        for line in f:
+            if not(sections in line):
+                op.write(line)
+        op.close()
     result = subprocess.call("echo '$$ %s EXECUTION START: %s' >> %s"
                             % (bench_name,
                                 str(starttime)[:19],
@@ -91,6 +101,7 @@ def run_all_cases(target_exec_dir, target, kind_bench, bench_name,
 	 tar_ip = settings.get_value('CLIENT', 'ip', type=str) 
 	 target.run("if [[ ! -e /mnt/ltp ]]; then mkdir -p /mnt/ltp; fi")
 # fix me , now that we create the folder, why not we mount it directly here
+
 	 s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
 	 try:
 # fix me , getting host ip to be optimised
@@ -681,17 +692,21 @@ def print_format():
     logging.info("="*55)
 
 
-def run_caliper_tests(target):
-    if os.path.exists(Folder.exec_dir):
-        shutil.rmtree(Folder.exec_dir)
-    os.mkdir(Folder.exec_dir)
+def run_caliper_tests(target,f_option):
+    #f_option =1 if -f is used
+    if f_option == 1:
+        if not os.path.exists(Folder.exec_dir):
+            os.mkdir(Folder.exec_dir)
+    else:
+        if os.path.exists(Folder.exec_dir):
+            shutil.rmtree(Folder.exec_dir)
+        os.mkdir(Folder.exec_dir)
     if not os.path.exists(Folder.results_dir):
         os.mkdir(Folder.results_dir)
     if not os.path.exists(Folder.yaml_dir):
         os.mkdir(Folder.yaml_dir)
     if not os.path.exists(Folder.html_dir):
         os.mkdir(Folder.html_dir)
-
     flag = 0
     target_execution_dir = server_utils.get_target_exec_dir(target)
     if not os.path.exists(target_execution_dir):
